chore: refactored cronjob executor.
